+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/modify_procedure_args_mode_original.sql
@@ -1,0 +1,8 @@
+CREATE OR REPLACE PROCEDURE public.get_user5(p_id_usder character, test integer)
+    LANGUAGE plpgsql
+    AS $$
+begin
+end; 
+$$;
+
+ALTER PROCEDURE public.get_user5(p_id_usder character, test integer) OWNER TO "SA";